requirements.txt
opencv-python-headless
onnxruntime
numpy
pillow
flask
flask-cors
flask-limiter
gunicorn